frontend/backend/requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
python-dotenv
aiosmtplib
pydantic
email-validator
